16-binary_tree_is_perfect.c

+++ 16-binary_tree_is_perfect.c
@@ -13,5 +13,6 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
 	height = binary_tree_height(tree);
 	size = binary_tree_size(tree);
 
-	return (binary_tree_is_full(tree) && size == (1 << height) - 1);
+    return (binary_tree_is_full(tree) && size == (size_t)((1 << height) - 1));
+
 }
